+++ controllers/contacts/getAllContacts.js
@@ -2,7 +2,10 @@ const  Contact  = require("../../models/contact");
 
 const getAllContacts = async (req, res) => {
   const { _id: owner } = req.user;
-  const result = await Contact.find({owner}).pupulate("owner","name email");
+  
+  const { page = 1, limit = 10 } = req.query;
+  const skip = (page - 1) * limit;
+  const result = await Contact.find({owner},"-email",{skip, limit});
   res.json(result);
 };
