+++ middlewares/authenticate.js
@@ -1,19 +1,22 @@
 const jwt = require("jsonwebtoken");
 
-const { User }  = require("../models/user");
+const  User  = require("../models/user");
 
 const { HttpError } = require("../helpers");
 
 const { SECRET_KEY }=process.env;
 
 const authenticate = async(req, res, next) => {
-    const { authorization= "" } = req.headers;
+    const { authorization = "" } = req.headers;
+    
     const [bearer, token] = authorization.split(" ");
+    
     if (bearer !== "Bearer") {
         next(HttpError(401));
     }
     try {
         const { id } = jwt.verify(token, SECRET_KEY);
+       
         const user = await User.findById(id);
         if (!user|| !user.token || user.token !== token) {
             next(HttpError(401));
