+++ models/user.js
@@ -1,6 +1,6 @@
 const { Schema, model } = require("mongoose");
-const Joi = require("joi");
-const { handleMongooseError } = require("../middlewares");
+
+const  handleMongooseError  = require("../middlewares/handleMongooseError");
 
 const emailRegexp = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
 
@@ -31,21 +31,6 @@ const userSchema = new Schema({
 
 userSchema.post("save", handleMongooseError);
 
-const registerSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().pattern(emailRegexp).required(),
-  password: Joi.string().min(6).required(),
-});
-
-const loginSchema = Joi.object({
-  email: Joi.string().pattern(emailRegexp).required(),
-  password: Joi.string().min(6).required(),
-});
-
-const schemas = {
-  registerSchema,
-  loginSchema,
-};
 const User = model("user", userSchema);
 
-module.exports = { schemas, User };
+module.exports = User ;
